LongTensor -> FloatTensor

+++ _03_homeworks/homework_2/titanic_dataset.py
@@ -10,7 +10,7 @@ pd.set_option("display.max_columns", None)
 class TitanicDataset(Dataset):
   def __init__(self, X, y):
     self.X = torch.FloatTensor(X)
-    self.y = torch.LongTensor(y)
+    self.y = torch.FloatTensor(y)
 
   def __len__(self):
     return len(self.X)
